Always store QR code URL in a file
It seems that the IARP API (RequestTransmittedDocs) does not return the
QR code URL along with the invoices. This means that this QR code can
only be retrieved from the success response by the API server,
when submitting an invoice.

Due to the importance of this field, we extend the 'send' operation with
a required argument, that makes the user specify where to save the
returned QR code URL.